+++ src/Akeneo/Pim/Enrichment/Bundle/Job/CleanRemovedAttribute/CleanProductTasklet.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Job\CleanRemovedAttribute;
+namespace Akeneo\Pim\Enrichment\Bundle\Job\CleanRemovedAttribute;
 
 use Akeneo\Pim\Enrichment\Bundle\Product\RemoveValuesFromProducts;
 use Akeneo\Pim\Enrichment\Component\Product\Query\CountProductsWithRemovedAttributeInterface;
