+++ src/Akeneo/Tool/Bundle/ConnectorBundle/Resources/translations/jsmessages.en_US.yml
@@ -12,6 +12,11 @@ batch_jobs:
         compute_family_variant_structure_changes: Compute family variant structure changes
         compute_completeness_of_products_family: Compute completeness of products family
         delete_products_and_product_models: Mass delete products
+        clean_attribute:
+            register_clean_job: Register the cleaning job
+            clean_product_model: Clean the removed attribute values in product models
+            clean_product: Clean the removed attribute values in products
+            whitelist_attribute_code: Whitelist the attribute code
     csv_product_export:
         label: Product export in CSV
         export.label: Product export
