+++ src/Oro/Bundle/PimDataGridBundle/Resources/translations/messages.en_US.yml
@@ -21,3 +21,4 @@ pim_datagrid:
       mass_delete: Mass Delete
       compute_family_variant_structure_changes: Compute family variant structure changes
       compute_completeness_of_products_family: Compute completeness of products family
+      clean_removed_attribute_job: Clean removed attribute values in products & product models
